requirements.txt
fastapi
uvicorn
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib